Tests moved to their own HealthCheck folder & iFate class renamed
